+++ client/src/components/Login,SignUp/Oauth/Oauth.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styles from './Oauth.module.css';
-import kakao from '../../images/kakao.png';
-import google from '../../images/google.png';
-import github from '../../images/github.png';
+import kakao from '../../../images/kakao.png';
+import google from '../../../images/google.png';
+import github from '../../../images/github.png';
 
 export default function Oauth() {
   return (
